Fix wrong hwaddr type

+++ projects/kea/fuzz_dhcp_pkt4.cc
@@ -97,7 +97,7 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
         pkt->getTransid();
         pkt->unpack();
         pkt->pack();
-        pkt->getMAC(fdp.ConsumeIntegral<uint32_t>());
+        pkt->getMAC(fdp.ConsumeIntegral<uint16_t>());
     } catch (...) {}
 
     // OptionVendor parsing
@@ -108,7 +108,7 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
             buf.begin(),
             buf.end()));
     }catch(...){}
-    
+
     try {
         // Package parsing for 4o6
         Pkt4Ptr pkt4 = Pkt4Ptr(new Pkt4(data, size));
@@ -119,7 +119,7 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
         pkt->getTransid();
         pkt->unpack();
         pkt->pack();
-        pkt->getMAC(fdp.ConsumeIntegral<uint32_t>());
+        pkt->getMAC(fdp.ConsumeIntegral<uint16_t>());
     } catch (...) {}
 
     try {
